pce_cd.cpp: Revert "move PCE CD emulation file into device/machine", Remove unused macro
