+++ src/mame/nec/pce_cd.cpp
