+++ src/mame/nec/pce_cd.h
